Simplify system uid display

+++ app/src/main/java/com/topjohnwu/magisk/ui/superuser/SuperuserViewModel.kt
@@ -3,6 +3,7 @@ package com.topjohnwu.magisk.ui.superuser
 import android.annotation.SuppressLint
 import android.content.pm.PackageManager
 import android.content.pm.PackageManager.MATCH_UNINSTALLED_PACKAGES
+import android.os.Process
 import androidx.databinding.ObservableArrayList
 import androidx.lifecycle.viewModelScope
 import com.topjohnwu.magisk.BR
@@ -59,10 +60,15 @@ class SuperuserViewModel(
             val policies = ArrayList<PolicyRvItem>()
             val pm = AppContext.packageManager
             for (policy in db.fetchAll()) {
-                val pkgs = pm.getPackagesForUid(policy.uid)
-                if (pkgs == null) {
-                    db.delete(policy.uid)
-                    continue
+                val pkgs: Array<String>?
+                if (policy.uid == Process.SYSTEM_UID) {
+                    pkgs = arrayOf("android")
+                } else {
+                    pkgs = pm.getPackagesForUid(policy.uid)
+                    if (pkgs == null) {
+                        db.delete(policy.uid)
+                        continue
+                    }
                 }
                 val map = pkgs.mapNotNull { pkg ->
                     try {
